graphrag/requirements.txt
langchain
neo4j
openai>=1.3.0
python-dotenv
tqdm
